moving file and added readme
